fix test_cpp

+++ include/txtvm/expr.h
@@ -48,7 +48,8 @@ public:
 };
 
 Expr IntConstant(int64_t value);
-Expr FloatConstant(int64_t value);
+// Expr FloatConstant(int64_t value);
+Expr FloatConstant(double value);
 
 class ExprNode : public Node {
 public:
